Add files via upload
The .m functions of the stochastic landscape classification algorithm: "Protein_State_Classification.m" and "give_vecs.m".
The first is the main function, and the latter is called by the first. 
The two data examples of the Chingolin and the Trp Cage are given, where the CVs are in "CVs_Chignolin.mat" and "CVs_Trp_Cage.mat", respectively, and their ground truth
are "GT_Chingolin.mat" and "GT_Tr_cage.mat", respectively.
For more details, please refer to the stochastic landscape classification upcoming paper!